pyproject.toml
[project]
name = "pk-exporter"
version = "0.1.0"
description = "Export PKNA dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "anthropic[bedrock]>=0.86.0",
    "docling-core>=2.25.0",
    "dspy>=3.1.3",
    "google-cloud-aiplatform>=1.143.0",
    "google-genai>=1.8.0",
    "jsonschema>=4.25.1",
    "jupyter>=1.1.1",
    "layoutparser[layoutmodels]>=0.3.4",
    "markdown-it-py>=4.0.0",
    "mlflow>=3.2.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic>=2.11.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
    "synthetic-data-kit>=0.0.5",
    "tiktoken>=0.11.0",
    "torch>=2.6.0",
    "transformers>=4.55.4",
    "unsloth>=2025.7.8",
    "vllm>=0.8.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.14.13",
    "ty>=0.0.12",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.uv]
link-mode = "symlink"
